replace double with float

+++ 01_qrcode/src/quirc_internal.h
@@ -52,7 +52,7 @@ struct quirc_capstone {
 
   struct quirc_point corners[4];
   struct quirc_point center;
-  double c[QUIRC_PERSPECTIVE_PARAMS];
+  float c[QUIRC_PERSPECTIVE_PARAMS];
 
   int qr_grid;
 };
@@ -72,7 +72,7 @@ struct quirc_grid {
 
   /* Grid size and perspective transform */
   int grid_size;
-  double c[QUIRC_PERSPECTIVE_PARAMS];
+  float c[QUIRC_PERSPECTIVE_PARAMS];
 };
 
 struct quirc {
